Suppress GitHub style user mentions
This commit adds support for providing a list of user mentions, via
user-mentions-to-escape.txt, to be suppressed (escaped) as part of the
migration in order to avoid sending alerts to whatever GH users they
happen to match.

To facilitate with the creation of the list, this commit also adds
InfoGatheringApp.java which helps to gather information from JIRA
about that can be used to prepare for the migration.