4th iteration
- Use WebClient to parallelize loading of votes and commits which is
2 extra calls per issue (or 35,000+ for SPR).

- Replace every SPR reference with a link.

- Initialize issues to import in a separate loop ahead of time, rather
than after each import, a) to minimize possibility of failure in the
middle of the import, and b) to ensure minimum processing time between
import requests for which we wait a minimum of 1 seconds.

- On every 100 imported issues, and at the end, check for failures and
write failures to a file, as well as all successfully imported issue
mappings to a separate file.