JiraClient is now fully concurrent
The issues are now also loaded concurrently, 5 pages at a time, which
loads in 30-35 secs vs 2-3 minutes.

The votes and commits which are 2 separate calls per issue are also
loaded concurrently with a variable for the level of concurrency.
A flatMap concurrency of 10, which is actually 20 concurrent HTTP calls,
gets through 35K+ calls for the Spring Framework under 10 min, vs over
an hour.